% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attractorPlot.R
\name{attractorDimplot}
\alias{attractorDimplot}
\title{Attractor Dimplot}
\usage{
attractorDimplot(seuratObj, attr.list, label = F, cols = NULL)
}
\arguments{
\item{seuratObj}{A Seurat object. For example,
\code{CASCC.output$res.predictK$adata}.}

\item{attr.list}{A list of attractors to plot. For example,
\code{CASCC.output$mainType.output$attrs.center}.}

\item{label}{Whether to label attractor on the plot.}
}
\description{
This function plot the CASCC UMAP reuslts. Each dot will be labeled with
the attractor it belongs to.
}
\seealso{
\code{\link{attractorPlot}}
}
\author{
Lingyi Cai
}
