% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseAttractorList.R
\name{attractorModify}
\alias{attractorModify}
\title{Modify Attractors}
\usage{
attractorModify(
  attrs,
  adata,
  expressionLevel = 0,
  percentage = 0.02,
  overlapN = 10
)
}
\arguments{
\item{attrs}{A list of attractors.}

\item{adata}{A Seurat object.}

\item{expressionLevel}{The expression level to be used to determine whether
an attractor is expressed in a cell.}

\item{percentage}{The percentage of cells that an attractor should express in.}

\item{overlapN}{The number of top features that are allowed to overlap
between two attractors.}
}
\description{
Remove attractors do not express in at least x% cells, by default we skip 
this step (percentage = 0%)
}
