% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attractorPlot.R
\name{attractorPlot}
\alias{attractorPlot}
\title{Attractor Plot}
\usage{
attractorPlot(seuratObj, attr.list, n.col = 4)
}
\arguments{
\item{seuratObj}{A Seurat object. For example,
\code{CASCC.output$res.predictK$adata}.}

\item{attr.list}{A list of attractors to plot. For example,
\code{CASCC.output$mainType.output$attrs.center},
\code{CASCC.output$fs.res$finalAttrs}.}

\item{n.col}{Specify the number of columns in which the plots should be
arranged when they are displayed together.}
}
\description{
This function colors singel cell on UMAP according to the expression levels
of co-expressed signatures.
}
\seealso{
\code{\link{attractorDimplot}}
}
\author{
Lingyi Cai
}
