% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASCC.featureSelection.R
\name{CASCC.featureSelection}
\alias{CASCC.featureSelection}
\title{CASCC Feature Selection}
\usage{
CASCC.featureSelection(
  data,
  inputDataType = "well-normalized",
  attr.raw = NULL,
  exponent.max = 10,
  exponent.min = 2,
  generalSeedList = NULL,
  mc.cores = NULL,
  topDEGs = 10,
  topAttr = 50,
  removeMT.RP.ERCC = TRUE,
  topN.DEG.as.seed = 1,
  removeNonProtein = FALSE,
  overlapN = 10,
  attractorModify.percentage = 0.01,
  min.diff.pct = 0.5,
  min.pct = 0.25,
  logfc.threshold = 0.25
)
}
\arguments{
\item{data}{Gene expression matrix.}

\item{inputDataType}{Input data type. Default is "well-normalized".
Options: "well-normalized", "needLog2" or "UMI". If the input matrix is
normalized and log-transformed, use "well-normalized". If the input matrix
is normalized but not log-transformed, use "needLog2". If the input matrix
is UMI-count, use "UMI".}

\item{attr.raw}{Raw attractor object. Default is NULL.  If users already
obtained attractor scanning results \code{CASCC.output$fs.res$attrs.raw},
they can provide the list of attractors to skip the scanning step.}

\item{exponent.max}{Maximum exponent for adaptive attractor. Default is 10.
To facilitate the analysis of large dataset,
users can set a smaller value, such as 5.}

\item{exponent.min}{Minimum exponent for adaptive attractor. Default is 2.}

\item{generalSeedList}{Seed genes set by users. Default is NULL. For example,
\code{generalSeedList = c("AMBP", "KRT19", "PRSS1", "CHGB", "RGS5", "LUM", "CDH5", "AIF1", "CD3D", "MS4A1")}.}

\item{mc.cores}{Number of cores to use.}

\item{topDEGs}{Number of top DEGs in each cluster selected as features.
Default is 10.}

\item{topAttr}{Number of top genes in each attractor selected as features.
Default is 50.}

\item{removeMT.RP.ERCC}{If remove mitochondrial genes, ribosomal genes and
spike-in genes from the analysis. Default is TRUE.}

\item{topN.DEG.as.seed}{The Number of top DEGs in each cluster to use as seeds. Default is
1.}

\item{removeNonProtein}{If remove non-protein coding genes.  Default is
FALSE.}

\item{overlapN}{Two attractors are considered duplicated if their top 50 genes
overlap by N genes. The default is 10.}
}
\value{
\item{features}{Selected features.} \item{adata}{Seurat object
after gene filtering.} \item{dfMarkers}{Differential expression genes for
initial clusters.} \item{markers}{Seeds for attractors.}
\item{attrs.raw}{Raw results from \code{link{scanSeeds.adaptive}.}}
\item{attrs}{A list of attractors in which identical attractors were
removed.} \item{finalAttrs}{A list of attractors in which duplicate and
identical attractors were removed.}
}
\description{
This function help select features for clustering.
}
\examples{
library("CASCC")
data("Data_PDAC_peng_2k")

# Step 0 - Step 3
fs.res <- CASCC.featureSelection(Data_PDAC_peng_2k, inputDataType = "well-normalized", attr.raw = NULL, exponent.max = 10, exponent.min = 2,
                                 generalSeedList = NULL, mc.cores = 1, topDEGs = 10, topAttr = 50,
                                 removeMT.RP.ERCC = TRUE, removeNonProtein = FALSE,
                                 topN.DEG.as.seed = 1,
                                 overlapN = 10)

}
\seealso{
\code{\link{filterGenes}}
}
