% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseAttractorList.R
\name{collapseAttractorList}
\alias{collapseAttractorList}
\title{Collapse AttractorList}
\usage{
collapseAttractorList(
  attractorList,
  NumTopFeature = 50,
  outputLength = NULL,
  verbose = FALSE,
  overlapN = 50,
  removeDominant = FALSE
)
}
\arguments{
\item{attractorList}{A list of attractors.}

\item{NumTopFeature}{Number of top features to be used in the attractors'
comparison.}

\item{verbose}{Whether to print out the progress.}

\item{overlapN}{The number of top features that are allowed to overlap
between two attractors.}
}
\description{
This function helps remove identical or duplicate attractors from a list of
attractors.
}
\author{
Lingyi Cai
}
