% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterGenes.R
\name{filterGenes}
\alias{filterGenes}
\title{Filter genes}
\usage{
filterGenes(
  data,
  min.cells = 0.05 * ncol(data),
  removeNonProtein = FALSE,
  removeMT.RP.ERCC = FALSE
)
}
\arguments{
\item{data}{Normalized data matrix.}

\item{min.cells}{default is 0. The genes that are not expressed in any
cells are removed.}

\item{removeNonProtein}{When \code{TRUE}, only keep protein coding genes.}

\item{removeMT.RP.ERCC}{When \code{TRUE}, filter out mitochondrial genes,
ribosomal genes, and spike-in gene.}
}
\description{
This function helps remove genes which are not relevant for characterizing
cell types.
}
\note{
First, the rows (features) with less than (min.cells + 1) cell that
have an expression level > 0 are excluded. Second, to remove genes not 
relevant for characterizing cell types, CASCC filtered out mitochondrial genes, 
ribosomal genes, and spike-in genes from the analysis by default.
}
\seealso{
\code{\link{CASCC.featureSelection}}
}
\author{
Lingyi Cai
}
