% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAttractor.adaptive.R
\name{findAttractor.2}
\alias{findAttractor.2}
\title{find Attractor}
\usage{
findAttractor.2(
  data,
  seed,
  a = 5,
  seedTopN = 100,
  maxIter = 100,
  epsilon = 1e-07,
  bin = 6,
  so = 3,
  rankBased = FALSE,
  negateMI = TRUE,
  verbose = TRUE
)
}
\description{
find Attractor
}
