% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanSeeds.adaptive.R
\name{findAttractor.adaptive_forMultiCores}
\alias{findAttractor.adaptive_forMultiCores}
\title{Scan seeds in adaptive manner with multiple cores}
\usage{
findAttractor.adaptive_forMultiCores(
  data,
  seed,
  epsilon = 1e-04,
  verbose = FALSE,
  exponent.max = 5,
  exponent.min = 2
)
}
\arguments{
\item{data}{A matrix of gene expression data. Rows are genes and columns
are cells.}

\item{seed}{Seed gene.}

\item{epsilon}{Threshold of convergence. By default, epsilon = 1e-4.}

\item{verbose}{When TRUE, it will show the top 20 genes of the metagene in
each iteration.}

\item{exponent.max}{max exponent value.}

\item{exponent.min}{min exponent value.}
}
\description{
Scan seeds in adaptive manner with multiple cores
}
\seealso{
\code{\link{findAttractor.adaptive}}
}
