% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASCC.R
\name{run.CASCC}
\alias{run.CASCC}
\title{Run CASCC in one line of code}
\usage{
run.CASCC(
  data.lognormcount,
  groundTruth.K = NULL,
  attr.raw = NULL,
  inputDataType = "well-normalized",
  mc.cores = 2,
  removeMT.RP.ERCC = TRUE,
  removeNonProtein = FALSE,
  topN.DEG.as.seed = 1,
  generalSeedList = NULL,
  exponent.max = 10,
  exponent.min = 2,
  topDEGs = 10,
  topAttr = 50,
  min.nc_fix = FALSE,
  overlapN = 10,
  attractorModify.percentage = 0
)
}
\arguments{
\item{data.lognormcount}{Input gene expression matrix.}

\item{groundTruth.K}{Default \code{NULL}. The number of clusters. By
default, CASCC will estimate the number of clusters. Users can manually set
the number of clusters if they prefer.}

\item{attr.raw}{Default \code{NULL}. If users already have attractor
scanning results \code{CASCC.output$fs.res$attrs.raw}, they can provide the
list of attractors.}

\item{inputDataType}{Default \code{"well-normalized"}.}

\item{mc.cores}{Default \code{NULL}. The number of cores to be used. By
default, CASCC use only 1 core.}

\item{exponent.max}{Default \code{10}.}

\item{exponent.min}{Default \code{2}.}

\item{topDEGs}{Default \code{10}. By default, the top 10 DEGs of each
cluster will be used as features in the following analysis.}

\item{topAttr}{Default \code{50}. By default, the top 50 genes of each
attractor will be used as features in the following analysis.}

\item{min.nc_fix}{Default \code{FALSE}. If \code{TRUE}, the minimum number
of clusters will be fixed to be 3. If \code{FALSE}, the minimum number of
clusters will be suggested by CASCC attractor output.}
}
\value{
\item{fs.res}{Output of \code{\link{CASCC.featureSelection}}. Step
1 to step 3 of CASCC.} \item{res.predictK}{Output of
\code{\link{predictClusterK}}. Step 4 of CASCC.} \item{res.kmeans}{Output
of run.kmeans function. Step 5.1 of CASCC. Kmeans with random centers.
}
\item{mainType.output}{Output of \code{\link{kmeansCenterModify}}. Step 5.2
of CASCC. Kmeans with fixed centers. }
}
\description{
Run the CASCC package with one line of code.
}
\author{
Lingyi Cai
}
