% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeansCenterModify.R
\name{run.kmeans.fixCenters}
\alias{run.kmeans.fixCenters}
\title{Kmeans with centers Using attractors as centers}
\usage{
run.kmeans.fixCenters(data.full, features, attrs)
}
\arguments{
\item{data.full}{The full data matrix}

\item{features}{The features}

\item{attrs}{The attractors}
}
\description{
Kmeans with centers Using attractors as centers
}
