% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeansCenterModify.R
\name{run.kmeans}
\alias{run.kmeans}
\title{Run kmeans}
\usage{
run.kmeans(data.lognormcount, centers)
}
\arguments{
\item{data.lognormcount}{The log normalized count matrix}

\item{centers}{The number of centers}
}
\description{
Run kmeans
}
