% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAttractor.adaptive.R
\name{findAttractor.adaptive}
\alias{findAttractor.adaptive}
\title{Find attractors using adaptive attractor finding algorithm}
\usage{
findAttractor.adaptive(
  data,
  seed,
  exponent.max = 5,
  exponent.min = 2,
  step.large = 1,
  step.small = 0.1,
  dominantThreshold = 0.2,
  seedTopN = 100,
  compareNth = 10,
  verbose = FALSE,
  epsilon = 1e-07,
  minimize = FALSE,
  first.idx = 1,
  second.idx = 2
)
}
\arguments{
\item{data}{The gene expression matrix.}

\item{seed}{The seed gene.}

\item{exponent.max}{The maximum exponent. Default 10.}

\item{exponent.min}{The minimum exponent. Default 2.}

\item{step.large}{Decreasing step of the first round of scanning.}

\item{step.small}{Decreasing step of the second round of scanning.}

\item{dominantThreshold}{The threshold for dominant genes.}

\item{seedTopN}{The threshold for divergency.}

\item{compareNth}{It will maxmize the weight of the top Nth gene.}

\item{verbose}{When TRUE, it will show the top 20 genes of the metagene in
each iteration.}

\item{epsilon}{The threshold for convergence.}

\item{minimize}{Default FALSE,when minimize = TRUE, the algorithm will
minize the difference of metagene[first.idx] and metagene[second.idx]
instead of maxmizing the weight of the top Nth gene.}

\item{first.idx}{Default first.idx = 1, the first gene of the metagene in
each iteration.}

\item{second.idx}{Default second.idx = 2, the second gene of the metagene
in each iteration.}
}
\description{
Find adaptive attractors using adaptive attractor finding algorithm. The
algorithm will scan the exponent from exponent.max to exponent.min with
step.large, and then scan the exponent from exponent.min to exponent.max
with scanning steps. The algorithm will maximize the strength of the
attractor.
}
\seealso{
\code{\link{scanSeeds.adaptive}}
}
\author{
Lingyi Cai
}
