% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictK.R
\name{predictClusterK}
\alias{predictClusterK}
\title{Predict the number of clusters}
\usage{
predictClusterK(
  data,
  fs.res,
  method = "ward.D2",
  index = "silhouette",
  min.nc_fix = TRUE,
  if_PCA_input = FALSE,
  nPCA = NULL
)
}
\arguments{
\item{data}{A matrix of gene expression data. Rows are genes and columns
are cells.}

\item{fs.res}{A list of feature selection results generated by
\code{CASCC.featureSelection} function.}

\item{method}{The cluster analysis method to be used. Default is "ward.D2".}

\item{index}{The index to be calculated. Default is "silhouette".}

\item{min.nc_fix}{Whether to fix the minimum number of clusters. Default is
FALSE.}

\item{if_PCA_input}{Whether to use PCA to predict the number of clusters. Default is
FALSE. Then use UMAP 2D embedding}

\item{nPCA}{The number of PCA. Default is NULL}
}
\value{
\item{clusterK}{The estimated number of clusters.}
\item{adata}{Seurat object after CASCC feature selection.}
}
\description{
Predict the number of clusters
}
\author{
Lingyi Cai
}
