% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASCC.R
\name{run.CASCC}
\alias{run.CASCC}
\title{Run CASCC in one line of code}
\usage{
run.CASCC(
  data.lognormcount,
  groundTruth.K = NULL,
  attr.raw = NULL,
  inputDataType = "well-normalized",
  mc.cores = 2,
  removeMT.RP.ERCC = TRUE,
  removeNonProtein = FALSE,
  topN.DEG.as.seed = 1,
  generalSeedList = NULL,
  exponent.max = 10,
  exponent.min = 2,
  topDEGs = 10,
  topAttr = 50,
  min.nc_fix = FALSE,
  if_PCA_input = TRUE,
  nPCA = 30,
  overlapN = 10,
  attractorModify.percentage = 0
)
}
\arguments{
\item{data.lognormcount}{Input gene expression matrix.}

\item{groundTruth.K}{Default \code{NULL}. The number of clusters. By
default, CASCC will estimate the number of clusters. Users can manually set
the number of clusters if they have prior knowledge about the number of clusters.}

\item{attr.raw}{Default \code{NULL}. If users already run CASCC and have attractor
scanning results \code{CASCC.output$fs.res$attrs.raw}, they can provide the
list of attractors.}

\item{inputDataType}{Input data type. Default is "well-normalized".
Options: "well-normalized", "needLog2" or "UMI". If the input matrix is
normalized and log-transformed, use "well-normalized". If the input matrix
is normalized but not log-transformed, use "needLog2". If the input matrix
is UMI-count, use "UMI".}

\item{mc.cores}{Default \code{NULL}. The number of cores to be used. By
default, CASCC use 2 cores.}

\item{removeMT.RP.ERCC}{Default \code{TRUE}. By default, mitochondrial genes,
ribosomal genes, and spike-in genes will be removed.}

\item{removeNonProtein}{Default \code{FALSE}. By default, non-protein coding
genes will not be removed.To facilitate the analysis, users can remove them.}

\item{topN.DEG.as.seed}{Default \code{1}. By default, the top 1 DEG of each
initial cluster will be used as seeds in the following analysis.}

\item{generalSeedList}{Seed genes set by users. Default is NULL. For example,
\code{generalSeedList = c("AMBP", "KRT19", "PRSS1", "CHGB", "RGS5", "LUM", "CDH5", "AIF1", "CD3D", "MS4A1")}.}

\item{exponent.max}{Default \code{10}.To facilitate the analysis of large dataset,
users can set a smaller value, such as 5.}

\item{exponent.min}{Default \code{2}.}

\item{topDEGs}{Default \code{10}. By default, the top 10 DEGs of each
initail cluster will be used as features in the following analysis.}

\item{topAttr}{Default \code{50}. By default, the top 50 genes of each
attractor will be used as features in the following analysis.}

\item{min.nc_fix}{Default \code{FALSE}. If \code{TRUE}, the minimum number
of clusters will be fixed to be 3. If \code{FALSE}, the minimum number of
clusters will be suggested by CASCC attractor output.}

\item{if_PCA_input}{Default \code{TRUE}. If \code{TRUE}, the input will be
PCA transformed. If \code{FALSE}, the input will be reduced to 2D UMAP
space. If \code{"feature-matrix"}, the input will be the feature matrix.}

\item{nPCA}{Default \code{30}. The number of PCA components to be used.}

\item{overlapN}{Default \code{10}. The number of overlapping genes between
different attractors to define duplicated attractors.}

\item{attractorModify.percentage}{Default \code{0}.
Remove attractor expressed in less than \code{attractorModify.percentage} percent of cells. This is disabled by default.}
}
\value{
\item{fs.res}{Output of \code{\link{CASCC.featureSelection}}. Step
1 to step 3 of CASCC.}
\item{res.predictK}{Output of
\code{\link{predictClusterK}}. Step 4 of CASCC.}
\item{mainType.output}{Output of \code{\link{kmeansCenterModify}}. Step 5
of CASCC. Kmeans with fixed centers. }
\item{res.kmeans}{Output
of run.kmeans function. Kmeans with random centers.
}
}
\description{
Run the CASCC package with one line of code.
}
\examples{
# load the toy dataset
data("Data_PDAC_peng_2k")
dim(Data_PDAC_peng_2k) # 400 cells and 2000 genes

# run CASCC
res <- CASCC::run.CASCC(Data_PDAC_peng_2k) # inputDataType `data` is normalized and log-transformed.

# clustering results
labels <- res$mainType.output$clusteringResults

}
