% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratProcessing.R
\name{SeuratProcessing}
\alias{SeuratProcessing}
\title{Intial clustering}
\usage{
SeuratProcessing(
  data,
  topN.DEG.as.seed = 1,
  resolution = 2,
  needNorm = FALSE,
  min.diff.pct = 0.5,
  topDEGs = 10
)
}
\arguments{
\item{data}{Gene expression matrix}

\item{topN.DEG.as.seed}{By default, topN = 1. TopN ranked differentially expressed
genes (DEGs) of each of the clusters found in initial clusters will be used
as the seed genes.}

\item{resolution}{By default, resolution = 2. Resolution parameter in
\code{Seurat::FindClusters}}

\item{needNorm}{If the input data needs normalization. When \code{TRUE},
the input matrix is UMI count, we use \code{Seurat::NormalizeData} to
normalize.}

\item{min.diff.pct}{\code{Seurat::FindAllMarkers.}}

\item{topDEGs}{The \code{topDEGs} top-ranked DEGs of each cluster will be
used as features in the following analysis.}
}
\description{
To prepare the list of seed genes, we apply the low-complexity "Seurat"
workflow to the gene expression matrix.
}
\examples{
library("CASCC")
data("Data_PDAC_peng_2k") 
data <- filterGenes(Data_PDAC_peng_2k)
seuratObject <- SeuratProcessing(data, topN.DEG.as.seed = 1, resolution = 2, needNorm = FALSE, min.diff.pct = 0.5, topDEGs = 10)
}
\seealso{
\code{\link{CASCC.featureSelection}}
}
\author{
Lingyi Cai
}
